+++ records/descriptors/DescriptorBase.cpp
@@ -20,7 +20,7 @@
 #include "StreamHelpers.h"
 
 
-void PropertyDescriptor::prefix(std::ostream& os, uint8_t indent) const
+void PropertyDescriptor::prefix(std::ostream& os, uint16_t indent) const
 {
     // Each property is preceded by its name and a colon. It is then up to the particular 
     // type of descriptor how to format the property's value.
