+++ records/descriptors/IndustryLayoutDescriptor.cpp
@@ -66,7 +66,7 @@ void IndustryTile::write(std::ostream& os) const
             write_uint16(os, tile);
             break;
         case Type::OldTile:
-            write_uint8(os, tile);
+            write_uint8(os, uint8_t(tile));
     }
 }
 
@@ -156,7 +156,7 @@ void IndustryLayouts::read(std::istream& is)
 
 void IndustryLayouts::write(std::ostream& os) const
 {
-    write_uint8(os, layouts.size());
+    write_uint8(os, uint8_t(layouts.size()));
 
     // Size needs to be calculated... 
     std::ostringstream ss;
@@ -165,7 +165,7 @@ void IndustryLayouts::write(std::ostream& os) const
         layout.write(ss);
     }
 
-    write_uint32(os, ss.str().length());
+    write_uint32(os, uint32_t(ss.str().length()));
     for (const auto& layout: layouts)
     {
         layout.write(os);
